# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminQuestionDto": {
        "properties": {
          "basePoints": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/AdminQuestionOptionDto"
            },
            "type": "array"
          },
          "stem": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stem",
          "explanation",
          "basePoints",
          "isActive",
          "createdBy",
          "createdAt",
          "updatedAt",
          "options"
        ],
        "type": "object"
      },
      "AdminQuestionOptionDto": {
        "properties": {
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "label": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "questionId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "questionId",
          "label",
          "content",
          "position",
          "isCorrect"
        ],
        "type": "object"
      },
...
      "ApiErrorDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "example": "请求参数验证失败",
            "type": "string"
          },
          "requestId": {
            "example": "7da2aa93-ef82-45c8-9df0-84232e6a5b13",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "details"
        ],
        "type": "object"
      },
...
      "CreateQuestionRequestDto": {
        "properties": {
          "basePoints": {
            "default": 10,
            "maximum": 1000,
            "minimum": 1,
            "type": "number"
          },
          "explanation": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "options": {
            "description": "标签与位置不可重复，且必须恰好有一个正确选项",
            "items": {
              "$ref": "#/components/schemas/QuestionOptionWriteRequestDto"
            },
            "maxItems": 6,
            "minItems": 2,
            "type": "array"
          },
          "stem": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "stem",
          "explanation",
          "basePoints",
          "options"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "QuestionListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AdminQuestionDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "QuestionOptionWriteRequestDto": {
        "properties": {
          "content": {
            "maxLength": 1000,
            "minLength": 1,
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "label": {
            "example": "A",
            "maxLength": 16,
            "minLength": 1,
            "type": "string"
          },
          "position": {
            "maximum": 5,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "label",
          "content",
          "position",
          "isCorrect"
        ],
        "type": "object"
      },
...
      "UpdateQuestionRequestDto": {
        "properties": {
          "basePoints": {
            "maximum": 1000,
            "minimum": 1,
            "type": "number"
          },
          "explanation": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "options": {
            "description": "标签与位置不可重复，且必须恰好有一个正确选项",
            "items": {
              "$ref": "#/components/schemas/QuestionOptionWriteRequestDto"
            },
            "maxItems": 6,
            "minItems": 2,
            "type": "array"
          },
          "stem": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Android 与其他无 Cookie 客户端使用的访问令牌",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Web 客户端访问令牌 Cookie",
        "in": "cookie",
        "name": "pq_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "英语答题、积分、商品兑换与订单管理开放接口",
    "title": "Point Quest API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/questions": {
      "get": {
        "operationId": "adminListQuestions",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "分页查询题库",
        "tags": [
          "管理端-题库"
        ]
      },
      "post": {
        "operationId": "adminCreateQuestion",
        "parameters": [
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "创建英语选择题",
        "tags": [
          "管理端-题库"
        ]
      }
    },
    "/api/v1/admin/questions/{questionId}": {
      "get": {
        "operationId": "adminGetQuestion",
        "parameters": [
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "获取题目详情",
        "tags": [
          "管理端-题库"
        ]
      },
      "patch": {
        "operationId": "adminUpdateQuestion",
        "parameters": [
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "更新或停用题目",
        "tags": [
          "管理端-题库"
        ]
      }
    },
...
  },
  "servers": [],
...
}